+++ src/Position/DirectionDown.cpp
@@ -9,6 +9,6 @@ dir_ptr DirectionDown::opposite(){
 	return dir_ptr(new DirectionUp());
 }
 
-void DirectionDown::move(Position& pos, int speed){
+void DirectionDown::move(Position& pos, float speed){
 	pos.moveVertical(speed * -1);
 }
