+++ src/Position/DirectionLeft.h
@@ -12,7 +12,7 @@ class DirectionLeft: public Direction{
 
 		dir_ptr opposite();
 
-		void move(Position& pos, int speed);
+		void move(Position& pos, float speed);
 };
 
 #endif
