+++ src/Position/DirectionRight.cpp
@@ -9,6 +9,6 @@ dir_ptr DirectionRight::opposite(){
 	return dir_ptr(new DirectionLeft());
 }
 
-void DirectionRight::move(Position& pos, int speed){
+void DirectionRight::move(Position& pos, float speed){
 	pos.moveHorizontal(speed);
 }
