+++ src/Position/Position.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include "Position.h"
 
-Position::Position(int x, int y): x(x), y(y){}
+Position::Position(float x, float y): x(x), y(y){}
 		
 Position::~Position(){}
 
@@ -9,16 +9,16 @@ bool Position::operator==(const Position& other){
 	return this->x == other.x && this->y == other.y;
 }
 
-void Position::moveHorizontal(int speed){
+void Position::moveHorizontal(float speed){
 	this->x += speed;
 }
 
-void Position::moveVertical(int speed){
+void Position::moveVertical(float speed){
 	this->y += speed;
 }
 
-int Position::distanceTo(const Position& other){
-	int x = pow(this->x - other.x, 2);
-	int y = pow(this->y - other.y, 2);
+float Position::distanceTo(const Position& other){
+	float x = pow(this->x - other.x, 2);
+	float y = pow(this->y - other.y, 2);
 	return sqrt(x + y);
 }
